requirements.txt
fastapi
uvicorn
pypi-json
uvicorn[standard]
pymongo
aioredis
motor
